keypad

+++ SmartHome/DeviceInfo.h
@@ -138,7 +138,7 @@ enum{
 #pragma mark - TV
 -(NSData *) nextProgram:(NSString *)deviceID;
 -(NSData *) previousProgram:(NSString *)deviceID;
--(NSData *) switchProgram:(uint8_t)program deviceID:(NSString *)deviceID; //切换台
+-(NSData *) switchProgram:(uint16_t)program deviceID:(NSString *)deviceID; //切换台
 -(NSData *) changeTVolume:(uint8_t)percent deviceID:(NSString *)deviceID;
 
 #pragma mark - DVD
